requirements.txt
graphiti-core
neo4j
python-dotenv
requests
beautifulsoup4